+++ StripeFinancialConnections/StripeFinancialConnections/Source/Web/FinancialConnectionsWebFlowViewController.swift
@@ -169,12 +169,13 @@ extension FinancialConnectionsWebFlowViewController {
                         if
                             let paymentMethodId = Self.extractValue(from: returnUrl, key: "payment_method_id")
                         {
-                            let instantDebitsLinkedBank = InstantDebitsLinkedBankImplementation(
+                            let instantDebitsLinkedBank = InstantDebitsLinkedBank(
                                 paymentMethodId: paymentMethodId,
                                 bankName: Self.extractValue(from: returnUrl, key: "bank_name")?
                                 // backend can return "+" instead of a more-common encoding of "%20" for spaces
                                     .replacingOccurrences(of: "+", with: " "),
-                                last4: Self.extractValue(from: returnUrl, key: "last4")
+                                last4: Self.extractValue(from: returnUrl, key: "last4"),
+                                linkMode: elementsSessionContext?.linkMode
                             )
                             self.notifyDelegateOfSuccess(result: .instantDebits(instantDebitsLinkedBank))
                         } else {